"notes" are now private data containers

+++ src/app/api/notes.py
@@ -1,11 +1,11 @@
 # ---------------------------------------------------------------------------------------------
 # This file contains the JSON endpoints for notes, handling the CRUD operations with the db 
 #
-from fastapi import APIRouter, HTTPException, Path, Depends
+from fastapi import APIRouter, HTTPException, Path, Depends, status
 
 from app.api import crud
 from app.api.models import NoteDB, NoteSchema
-from app.api.users import User, get_current_active_user
+from app.api.users import User, get_current_active_user, user_has_role
 
 from typing import List
 
@@ -20,30 +20,43 @@ router = APIRouter()
 async def create_note(payload: NoteSchema, 
                       current_user: User = Depends(get_current_active_user)):
     
-    note_id = await crud.post_note(payload)
+    note_id = await crud.post_note(payload, current_user.id)
 
     response_object = {
         "id": note_id,
+        "owner": current_user.id,
         "title": payload.title,
         "description": payload.description,
+        "data": payload.data,
     }
     return response_object
 
 # ----------------------------------------------------------------------------------------------
 # Note: id's type is validated as greater than 0  
 @router.get("/{id}/", response_model=NoteDB)
-async def read_note(id: int = Path(..., gt=0),):
+async def read_note(id: int = Path(..., gt=0),
+                    current_user: User = Depends(get_current_active_user)):
     
     note = await crud.get_note(id)
     if not note:
         raise HTTPException(status_code=404, detail="Note not found")
-    return note
+    
+    if user_has_role( current_user, "admin") or current_user.id == note.owner:
+        return note
+    
+    raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, 
+                        detail="Not Authorized to access other's notes")
 
 # ----------------------------------------------------------------------------------------------
 # The response_model is a List with a NoteDB subtype. See import of List top of file. 
 @router.get("/", response_model=List[NoteDB])
-async def read_all_notes():
-    return await crud.get_all_notes()
+async def read_all_notes(current_user: User = Depends(get_current_active_user)):
+    if user_has_role( current_user, "admin"):
+        return await crud.get_all_notes()
+    
+    else:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, 
+                            detail="Not Authorized to access all notes")
 
 # ----------------------------------------------------------------------------------------------
 # Note: id's type is validated as greater than 0  
@@ -55,12 +68,18 @@ async def update_note(payload: NoteSchema, id: int = Path(..., gt=0),
     if not note:
         raise HTTPException(status_code=404, detail="Note not found")
 
+    if note.owner != current_user.id and not user_has_role(current_user,"admin"):
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, 
+                        detail="Not Authorized to change other's notes")
+        
     note_id = await crud.put_note(id, payload)
 
     response_object = {
         "id": note_id,
+        "owner": note.owner,
         "title": payload.title,
         "description": payload.description,
+        "data": payload.data,
     }
     return response_object
 
@@ -74,6 +93,10 @@ async def delete_note(id: int = Path(..., gt=0),
     if not note:
         raise HTTPException(status_code=404, detail="Note not found")
 
+    if note.owner != current_user.id and not user_has_role(current_user,"admin"):
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, 
+                        detail="Not Authorized to delete other's notes")
+        
     await crud.delete_note(id)
 
     return note
